all models implemented. Bug in Mov_Avg.cpp but i really dont know where

+++ src/t_models/CSum.cpp
@@ -83,12 +83,12 @@ void CSum::print(bool verbose)
 {
 }
 
-float CSum::estimate(uint32_t time)
+double CSum::estimate(uint32_t time)
 {
 	return score;
 }
 
-float CSum::predict(uint32_t time)
+double CSum::predict(uint32_t time)
 {
 	return score;
 }
